:sparkles: :tractor: :art: Refactor fields filters settings to use declarative flags
