+++ pkg/cli/flags/select.yaml
@@ -1,0 +1,10 @@
+- name: select
+  type: string
+  help: Select a single field and output as a single line
+  default: ""
+
+- name: select-template
+  type: string
+  help: Output a single templated value for each row, on a single line
+  default: ""
+
